Move back button to components
