+++ pages/solutions/[slug].tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { BeWithUs } from '../../blocks/be-with-us'
 import { Footer } from '../../blocks/footer'
 import { Header } from '../../blocks/header'
-import { BackButton } from '../../blocks/back-button'
+import { BackButton } from '../../components/back-button'
 import { Container } from '../../components/container'
 import { Heading } from '../../components/heading'
 import { useRouter } from 'next/router'
